Move favorite app out of the apps directory.
