+++ mapstory/favorite/api.py
