+++ mapstory/favorite/models.py
