+++ mapstory/favorite/templates/favorite/_favorite.html
