+++ mapstory/favorite/templates/favorite/_favorite_js.html
