+++ mapstory/favorite/templates/favorite/favorite_list.html
