+++ mapstory/favorite/utils.py
