+++ mapstory/favorite/views.py
