+++ mapstory/urls.py
@@ -22,7 +22,7 @@ from tastypie.api import Api
 from mapstory.api.api import MapstoryOwnersResource, InterestsResource
 from mapstory.api.resourcebase_api import ResourceBaseResource
 from mapstory.api.urls import api as mapstory_api
-from mapstory.apps.favorite.urls import api as favorites_api
+from mapstory.favorite.urls import api as favorites_api
 from mapstory.importers import UploadedLayerResource
 from mapstory.views import download_append_csv, download_append_shp
 from mapstory.views import GetPageView
@@ -104,7 +104,7 @@ urlpatterns = patterns('',
 
                        url(r'^donate$', LeaderListView.as_view(template_name='mapstory/donate.html'), name='donate'),
                        url(r'^proxy/', proxy),
-                       url(r'^favorite/', include('mapstory.apps.favorite.urls')),
+                       url(r'^favorite/', include('mapstory.favorite.urls')),
                        url(r"^flag/", include('mapstory.apps.flag.urls')),
 
                        # Layers
